requirements.txt
discord.py>=2.1.0,<3.0.0
